pyproject.toml
[project]
name = "new-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "langchain-community>=0.4.1",
    "langchain-fireworks>=1.1.0",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.2",
    "pypdf>=6.9.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
